bulk-mailer-system/worker-service/requirements.txt
confluent-kafka==2.3.0
python-dotenv
